Separate build and installation directory for HDF5 and adftool.
There are 2 separate build profiles for HDF5:
- build-hdf5-with-node
- build-hdf5-emscripten

The first one is just there to generate H5lib_settings.c, the second
one is the true thing. It installs in public/emscripten.

We build adftool in-place with the native compiler to run tests.

Then, there is 1 separate build profile for adftool:
build-adftool-emscripten, which also installs in public/emscripten.