Implement allocate and store for the HDF5 back-end.
So this is my first experience writing a program (that works) with
HDF5. I am pleased to find similarities between HDF5 and MPI, because
we first define the datatype that we want to map to the foreign
dataset (the dataspace in HDF5), a compatible datatype that we want to
map to our memory buffer (not necessarily equal, but with the same
number of elements), and then we do the read/write operation. A
difference is that while MPI read/write has a number of elements and a
complex type, HDF5 read/write has a base type and a complex
arrangement of elements. Of course HDF5 lets you create custom types,
but there’s no real point in that for ADF.

What is not clear in the HDF5 library however is that the arguments
are:

H5Dwrite/read (<DATASET>, <base type>, <MEMORY dataspace>, <DATASET
dataspace>, <properties>, <MEMORY>)

So we see that the dataset and memory arguments are not next to their
respective dataspaces in the call order. Since it is the same type,
the compiler can’t generate a warning if I do the wrong thing. And oh
boy I did do the wrong thing and it was not pleasant debugging those
invalid reads/writes.

For the dataspace, I first went with create_simple and then offset
them with offset_simple, but it seemed like the offset wouldn’t offset
anything to the dataspace. Weird. The recommended way is to use
select_hyperslab though, so I ended up with that and it works.

Before doing all the debugging, I tried to create an extensible
dataset in the accompanying example, but if you do that, since it is a
matrix, the dataset must be chunked. That’s a bit awkward, since the
matrix will ever grow in one dimension, so it is equivalent to
creating a 1-dimensional array of a 2*order+1-integers type, but the
ADF specification insists on a 2D matrix. Then I ran into all the
other problems, so I figured going for a non-extensible dataset in the
example would be better.

A final note about me using HDF5 is realizing how many things might go
wrong. Every function may fail in HDF5, but it’s not clear to me why
they might fail. I guess memory exhaustion errors. That means a lot to
clean up, so goto and labels are really helpful here.

In any case, in my API I decided to not allow the write and the
allocation to fail. In a more polished version of the HDF5 bindings, I
could set an error flag in the parameters object and refuse to do
additional things, and then adding a function to query the error
flag. But since there is no transaction support in HDF5, I don’t see
an example where the user knowing the update was partial could be able
to do anything about it.