Do not let the user change the channel identifier.
Many things depend on the channel identifier, such as the scale/offset
encoder and decoder for the raw data, and the type. The user may add
even more metadata. It would be expected from changing the channel
identifier that the scale/offset encoder/decoder also reflect that
change, so that the encoder/decoder still has the same values after
the rename. However, if I guarantee that the encoder/decoder will
change, I will have to do that for the types also. But what about the
other user’s metadata? Maybe I will have to change nearly every quad
in the file.

So, I choose not to let the user change the channel identifier.