Add a sample PKGBUILD
The easiest way to get adftool for windows is to cross-compile it. The
second easiest way is to build it with msys2. msys2 is built around
pacman, so here is a sample definition. Put this file next to a full
distribution source of adftool, change pkgver, hash the archive file
with sha256sum and replace the digest in the build file. Rename it to
PKGBUILD, and run:

    makepkg-mingw --cleanbuild --syncdeps --force --noconfirm

* PKGBUILD.example: New file.